+++ src/rcpp_bcDist_str.cpp
@@ -78,12 +78,13 @@ int rcpp_str_dist_led(String x, String y) {
 //' @noRd
 // [[Rcpp::export(.rcpp_bcDist_str_v)]]
 SEXP rcpp_bcDist_str_v(
-  CharacterVector x, CharacterVector y, 
+  SEXP x, SEXP y, 
   R_xlen_t nout, int op
 ) {
 
 
-
+const SEXP *px = STRING_PTR_RO(x);
+const SEXP *py = STRING_PTR_RO(y);
 
 SEXP out = PROTECT(Rf_allocVector(INTSXP, nout));
 int *pout;
@@ -104,12 +105,13 @@ return out;
 //' @noRd
 // [[Rcpp::export(.rcpp_bcDist_str_ov)]]
 SEXP rcpp_bcDist_str_ov(
-  CharacterVector x, CharacterVector y,  bool RxC, SEXP out_dim,
+  SEXP x, SEXP y,  bool RxC, SEXP out_dim,
   R_xlen_t nout, int op
 ) {
 
 
-
+const SEXP *px = STRING_PTR_RO(x);
+const SEXP *py = STRING_PTR_RO(y);
 
 SEXP out = PROTECT(Rf_allocVector(INTSXP, nout));
 int *pout;
@@ -129,13 +131,14 @@ return out;
 //' @noRd
 // [[Rcpp::export(.rcpp_bcDist_str_d)]]
 SEXP rcpp_bcDist_str_d(
-  CharacterVector x, CharacterVector y, 
+  SEXP x, SEXP y, 
   SEXP by_x,
   SEXP by_y,
   SEXP dcp_x, SEXP dcp_y, SEXP out_dim, R_xlen_t nout, int op
 ) {
 
-
+const SEXP *px = STRING_PTR_RO(x);
+const SEXP *py = STRING_PTR_RO(y);
 
 SEXP out = PROTECT(Rf_allocVector(INTSXP, nout));
 int *pout;